#Fix propblem with db connection string

+++ TechnicoRMP.Shared/Dtos/UpdatePropertyRepair.cs
@@ -8,7 +8,7 @@ public class UpdatePropertyRepair : IDto
     public EnTypeOfRepair? TypeOfRepair { get; set; }
     public string? Address { get; set; }
     public EnRepairStatus? RepairStatus { get; set; }
-    public decimal? Cost { get; set; }
+    public double? Cost { get; set; }
     public long? RepairerId { get; set; }
     public bool? IsActive { get; set; } = true;
 }
